Changes due to moving the loader classes into TwigLoader directory
