+++ lib/TwigLoader/CallerCompatibleInterface.php
@@ -1,9 +1,13 @@
 <?php
 
-namespace Timber;
+namespace Timber\TwigLoader;
 
-
-interface CallerCompatibleLoaderInterface
+/**
+ * @author Heino H. Gehlsen <heino@gehlsen.dk>
+ * @copyright 2017 Heino H. Gehlsen
+ * @license MIT
+ */
+interface CallerCompatibleInterface
 {
 	public function setCaller($caller = false);
